% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{area_by_class}
\alias{area_by_class}
\title{Area by burn severity}
\usage{
area_by_class(x, unit = "ha")
}
\arguments{
\item{x}{A classified burn severity raster returned by
\code{classify_dnbr()}.}

\item{unit}{Area units. One of \code{"ha"}, \code{"km2"} or \code{"m2"}.}
}
\value{
A data frame containing class, label, area and percentage.
}
\description{
Calculate the area occupied by each burn severity class.
}
