% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_band.R
\name{build_band}
\alias{build_band}
\title{Build bands}
\usage{
build_band(collection, asset, scl_stacks = NULL)
}
\arguments{
\item{collection}{An sbr_collection.}

\item{assets}{Assets to include.}
}
\value{
A SpatRaster.
}
\description{
Build bands
}
