% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{burn_caption}
\alias{burn_caption}
\title{Create a caption describing a burn analysis}
\usage{
burn_caption(x)
}
\arguments{
\item{x}{A burn analysis object.}
}
\value{
A character string suitable for use as a plot caption.
}
\description{
Create a caption describing a burn analysis
}
