% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_summary.R
\name{burn_summary}
\alias{burn_summary}
\title{Summarise burn statistics}
\usage{
burn_summary(burn, boundary = NULL, id = NULL)
}
\arguments{
\item{burn}{An \code{sbr_burn} object.}

\item{boundary}{Optional boundary. If omitted, \code{burn$boundary} is used.}

\item{id}{Optional attribute used to name polygons.}
}
\value{
An object of class \code{sbr_summary}.
}
\description{
Summarise burn statistics
}
