% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{calc_humidity}
\alias{calc_humidity}
\title{Calculate relative humidity and vapour pressure deficit}
\usage{
calc_humidity(temperature_c, dewpoint_c)
}
\arguments{
\item{temperature_c}{Air temperature in degrees Celsius.}

\item{dewpoint_c}{Dew-point temperature in degrees Celsius.}
}
\value{
A data frame containing \code{relative_humidity}
in percent and \code{vpd_kpa} in kPa.
}
\description{
Calculate relative humidity and vapour pressure deficit
}
