% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calc_ndmi}
\alias{calc_ndmi}
\title{Calculate NDMI}
\usage{
calc_ndmi(x)
}
\arguments{
\item{x}{A SpatRaster containing \code{nir08} and \code{swir22}.}
}
\value{
A single-layer SpatRaster named \code{ndmi}.
}
\description{
Calculates the Normalized Difference Moisture Index.
}
