% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calc_ndvi}
\alias{calc_ndvi}
\title{Calculate NDVI}
\usage{
calc_ndvi(x)
}
\arguments{
\item{x}{A SpatRaster containing \code{nir08} and \code{red}.}
}
\value{
A single-layer SpatRaster named \code{ndvi}.
}
\description{
Calculates the Normalized Difference Vegetation Index.
}
