% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-utils.R
\name{calc_wind}
\alias{calc_wind}
\title{Calculate wind speed and direction}
\usage{
calc_wind(u_ms, v_ms)
}
\arguments{
\item{u_ms}{Numeric vector of eastward wind components in m/s.}

\item{v_ms}{Numeric vector of northward wind components in m/s.}
}
\value{
A data frame containing:
\describe{
\item{wind_speed_ms}{Wind speed in m/s.}
\item{wind_direction_deg}{Meteorological direction from which
the wind is blowing, in degrees clockwise from north.}
}
}
\description{
Calculates wind speed and meteorological wind direction
from eastward (u) and northward (v) wind components.
}
