% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create a sentinelBurnR project}
\usage{
create_project(name, root = getOption("sbr.project_dir"))
}
\arguments{
\item{name}{Project name.}

\item{root}{Root project directory. If NULL, uses the
\code{sbr.project_dir} option.}
}
\value{
An sbr_project object.
}
\description{
Create a sentinelBurnR project
}
