% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{extract_rainfall}
\alias{extract_rainfall}
\title{Extract rainfall time series}
\usage{
extract_rainfall(climate, boundary, fun = mean)
}
\arguments{
\item{climate}{A SpatRaster returned by \code{read_climate()}.}

\item{boundary}{Boundary polygon.}

\item{fun}{Summary function used when extracting rainfall.}
}
\value{
An object of class \code{sbr_rainfall}.
}
\description{
Extract the mean daily precipitation for a boundary from an ERA5
precipitation raster.
}
