% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-collection.R
\name{files.sbr_collection}
\alias{files.sbr_collection}
\title{Access collection files}
\usage{
\method{files}{sbr_collection}(x, ...)
}
\arguments{
\item{x}{An \code{sbr_collection}.}

\item{...}{Not used.}
}
\description{
Access collection files
}
