% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{geometry}
\alias{geometry}
\title{Return the geometry from an AOI}
\usage{
geometry(x)
}
\arguments{
\item{x}{An sbr_aoi object.}
}
\value{
A terra SpatVector.
}
\description{
Return the geometry from an AOI
}
