% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_metadata.R
\name{index_info}
\alias{index_info}
\title{Spectral index metadata}
\format{
A named list.
}
\usage{
index_info
}
\description{
Internal metadata defining plotting and display properties for
supported spectral indices.
}
\keyword{internal}
