% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gallery.R
\name{palette_gallery}
\alias{palette_gallery}
\title{Display SentinelBurnR colour palettes}
\usage{
palette_gallery()
}
\value{
Invisibly returns a named list of colour vectors.
}
\description{
Displays the colour palettes available for plotting
SentinelBurnR results.
}
