% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel_apply}
\alias{parallel_apply}
\title{Run a function sequentially or in parallel
parallel_apply}
\usage{
parallel_apply(X, FUN, workers = 1, ...)
}
\arguments{
\item{X}{Input vector.}

\item{FUN}{Function to apply.}

\item{workers}{Number of workers.}
}
\description{
Run a function sequentially or in parallel
parallel_apply
}
\keyword{internal}
