% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sbr_aoi}
\alias{plot.sbr_aoi}
\title{Plot an Area of Interest}
\usage{
\method{plot}{sbr_aoi}(x, ...)
}
\arguments{
\item{x}{An sbr_aoi object.}

\item{...}{Additional arguments passed to terra::plot().}
}
\description{
Plot an Area of Interest
}
