% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_climate}
\alias{plot_climate}
\title{Plot climate conditions}
\usage{
plot_climate(x, metric = c("rainfall", "temperature"))
}
\arguments{
\item{x}{An \code{sbr_climate} object returned by
\code{\link[=analyse_climate]{analyse_climate()}}.}

\item{metric}{Climate metric to plot. One of \code{"rainfall"}
or \code{"temperature"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot rainfall or temperature conditions from an
\code{sbr_climate} analysis.
}
