% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_drought}
\alias{plot_drought}
\title{Plot drought analysis}
\usage{
plot_drought(
  x,
  index = c("anomaly", "standardised"),
  boundary = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{x}{An object of class \code{sbr_drought}.}

\item{index}{Drought product to plot. Either \code{"anomaly"} or
\code{"standardised"}.}

\item{boundary}{Optional boundary to overlay.}

\item{title}{Optional plot title.}

\item{subtitle}{Optional plot subtitle.}

\item{caption}{Optional plot caption.}
}
\value{
A ggplot object.
}
\description{
Plot vegetation moisture anomalies from an \code{sbr_drought}
analysis.
}
