% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fire_weather}
\alias{plot_fire_weather}
\title{Plot fire-weather conditions}
\usage{
plot_fire_weather(x)
}
\arguments{
\item{x}{An \code{sbr_fire_weather} object.}
}
\value{
A patchwork object containing fire-weather plots.
}
\description{
Plot fire-weather conditions
}
