% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scl.R
\name{plot_scl}
\alias{plot_scl}
\title{Plot Sentinel-2 Scene Classification Layer}
\usage{
plot_scl(collection, tile = 1, layer = 1)
}
\arguments{
\item{collection}{sbr_collection.}

\item{tile}{Tile index.}

\item{layer}{Date/layer index.}

\item{overlay}{Draw cloud classes on top of an RGB image?}
}
\description{
Plot Sentinel-2 Scene Classification Layer
}
