% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_climate.R
\name{print.sbr_climate}
\alias{print.sbr_climate}
\title{Print climate analysis}
\usage{
\method{print}{sbr_climate}(x, ...)
}
\arguments{
\item{x}{An \code{sbr_climate} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print climate analysis
}
