% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-landscape.R
\name{print.sbr_landscape}
\alias{print.sbr_landscape}
\title{Print a landscape analysis}
\usage{
\method{print}{sbr_landscape}(x, ...)
}
\arguments{
\item{x}{An \code{sbr_landscape} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a landscape analysis
}
