% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{read_band}
\alias{read_band}
\title{Read one asset from a downloaded Sentinel-2 collection}
\usage{
read_band(collection, asset)
}
\arguments{
\item{collection}{An sbr_collection.}

\item{asset}{Sentinel-2 asset name.}
}
\value{
A named list of SpatRaster objects, one per tile.
}
\description{
Returns one SpatRaster stack per MGRS tile.
}
