% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{read_boundary.sbr_aoi}
\alias{read_boundary.sbr_aoi}
\title{Read boundary from an sbr_aoi}
\usage{
\method{read_boundary}{sbr_aoi}(x, ...)
}
\arguments{
\item{x}{An \code{sbr_aoi} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
A \code{SpatVector}.
}
\description{
Read boundary from an sbr_aoi
}
