% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{read_climate}
\alias{read_climate}
\title{Read climate data}
\usage{
read_climate(files)
}
\arguments{
\item{files}{Path to one or more NetCDF files.}
}
\value{
A SpatRaster.
}
\description{
Read a climate raster from a NetCDF file.
}
