% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_visual.R
\name{read_visual}
\alias{read_visual}
\title{Read Sentinel-2 true-colour imagery}
\usage{
read_visual(collection, aoi, date = NULL)
}
\arguments{
\item{collection}{An \code{sbr_collection} containing the \code{visual} asset.}

\item{aoi}{Area of interest}

\item{date}{Date of the acquisition to read. If the collection contains
only one acquisition, this may be omitted.}
}
\value{
A three-layer \code{terra::SpatRaster} containing red, green and blue
true-colour imagery.
}
\description{
Reads the Sentinel-2 \code{visual} asset for a single acquisition,
crops the imagery to the collection AOI, and mosaics multiple
Sentinel-2 tiles where necessary.
}
