% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-raster.R
\name{save_raster}
\alias{save_raster}
\title{Save a raster to a project}
\usage{
save_raster(x, filename, project, folder)
}
\arguments{
\item{x}{A SpatRaster.}

\item{filename}{Output filename.}

\item{project}{sbr_project.}

\item{folder}{Project folder.}
}
\value{
Invisible filename.
}
\description{
Save a raster to a project
}
\keyword{internal}
