% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{sbr_options}
\alias{sbr_options}
\title{Configure sentinelBurnR}
\usage{
sbr_options(project_dir = NULL, cache_dir = NULL, temp_dir = NULL)
}
\arguments{
\item{project_dir}{Project directory}

\item{cache_dir}{cache directory}

\item{temp_dir}{temporary directory}
}
\description{
Configure sentinelBurnR
}
