% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_summary.R, R/summary.R
\name{summarise_polygon}
\alias{summarise_polygon}
\title{Summarise burn statistics within boundaries}
\usage{
summarise_polygon(burn, polygon)

summarise_polygon(burn, polygon)
}
\arguments{
\item{burn}{An sbr_burn object.}

\item{boundary}{Boundary polygons.}

\item{id}{Optional attribute giving polygon names.}
}
\value{
An sbr_summary object.

An sbr_summary object.
}
\description{
Summarise burn statistics within boundaries

Summarise burn statistics within boundaries
}
