% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-collection.R
\name{summary.sbr_collection}
\alias{summary.sbr_collection}
\title{Summarise a downloaded Sentinel-2 collection}
\usage{
\method{summary}{sbr_collection}(object, ...)
}
\arguments{
\item{object}{An sbr_collection object.}

\item{...}{Additional arguments (unused).}
}
\value{
A data frame describing the downloaded files.
}
\description{
Summarise a downloaded Sentinel-2 collection
}
