% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries_change}
\alias{timeseries_change}
\title{Calculate change in a spectral index through time}
\usage{
timeseries_change(x)
}
\arguments{
\item{x}{A data frame returned by \code{sentinel_timeseries()} or
\code{index_timeseries()}.}
}
\value{
A data frame containing the start and end dates of each
interval, the median index values and the change between them.
}
\description{
Calculates the change in median spectral index between consecutive
observations in a time series.
}
