% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_vegetation.R
\name{analyse_vegetation}
\alias{analyse_vegetation}
\title{Analyse vegetation condition}
\usage{
analyse_vegetation(collection, assets = s2_vegetation_assets)
}
\arguments{
\item{collection}{An \code{sbr_collection}.}

\item{assets}{Sentinel-2 assets used to build the composite.}
}
\value{
An object of class \code{sbr_vegetation}.
}
\description{
Build a Sentinel-2 composite and calculate vegetation
condition indices.
}
