% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{build_composite}
\alias{build_composite}
\title{Build a median Sentinel-2 composite}
\usage{
build_composite(
  collection,
  assets = s2_burn_assets,
  cache = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{collection}{An sbr_collection.}

\item{assets}{Assets to include.}
}
\value{
A SpatRaster.
}
\description{
Build a median Sentinel-2 composite
}
