% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{burn_area}
\alias{burn_area}
\title{Creates a burned / not-burned raster from a dNBR raster.}
\usage{
burn_area(x, unit = "ha")
}
\arguments{
\item{x}{A single-layer dNBR SpatRaster.}

\item{unit}{Area units to return (e.g. "ha", "m2", "km2").}
}
\value{
A single-layer SpatRaster named \code{burned}.
}
\description{
Creates a burned / not-burned raster from a dNBR raster.
}
