% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn.R
\name{calc_dnbr}
\alias{calc_dnbr}
\title{Calculate differenced Normalized Burn Ratio}
\usage{
calc_dnbr(pre, post)
}
\arguments{
\item{pre}{Pre-fire NBR raster.}

\item{post}{Post-fire NBR raster.}
}
\value{
A SpatRaster.
}
\description{
Calculate differenced Normalized Burn Ratio
}
