% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{.normalised_difference}
\alias{.normalised_difference}
\title{Calculates Normalized Differences}
\usage{
.normalised_difference(x, band1, band2, name)
}
\arguments{
\item{x}{A \code{SpatRaster} containing \code{band1} and \code{band2}.}
}
\value{
A single-layer \code{SpatRaster} named \code{ndvi}.
}
\description{
Calculates Normalized Differences
}
