% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_s2}
\alias{download_s2}
\title{Download Sentinel-2 assets}
\usage{
download_s2(
  x,
  assets = NULL,
  limit = NULL,
  max_cloud = NULL,
  output_dir = tools::R_user_dir("sentinelBurnR", which = "cache"),
  overwrite = FALSE,
  workers = 4
)
}
\arguments{
\item{x}{An \code{sbr_search} object.}

\item{assets}{Character vector of asset names.}

\item{limit}{Optional maximum number of scenes to download.}

\item{output_dir}{Directory used to cache downloaded files.}

\item{overwrite}{Logical. Overwrite existing files?}

\item{workers}{Number of parallel download workers.}
}
\value{
An \code{sbr_collection} object.
}
\description{
Download selected Sentinel-2 assets from an \code{sbr_search} object.
Existing cached files are skipped unless \code{overwrite = TRUE}.
}
