% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_s2}
\alias{download_s2}
\title{Download Sentinel-2 assets}
\usage{
download_s2(
  x,
  assets = s2_default_assets,
  limit = NULL,
  output_dir = tools::R_user_dir("sentinelBurnR", which = "cache"),
  overwrite = FALSE
)
}
\arguments{
\item{x}{An \code{sbr_search} object.}

\item{assets}{Character vector of asset names.}

\item{limit}{Optional maximum number of scenes to download.}

\item{output_dir}{Directory for downloaded files.}

\item{overwrite}{Should existing files be overwritten?}
}
\value{
An \code{sbr_collection} object.
}
\description{
Download selected assets from scenes returned by \code{search_s2()}.
}
