% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{get_rainfall}
\alias{get_rainfall}
\title{Retrieve rainfall data}
\usage{
get_rainfall(boundary, start, end, project = NULL)
}
\arguments{
\item{boundary}{Boundary polygon.}

\item{start}{Start date.}

\item{end}{End date.}

\item{project}{Optional sbr_project.}
}
\value{
A data.frame.
}
\description{
Retrieves daily rainfall for a boundary and date range.
}
