% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dnbr}
\alias{plot_dnbr}
\title{Plot differenced Normalized Burn Ratio}
\usage{
plot_dnbr(x, title = "Differenced Normalized Burn Ratio", subtitle = NULL)
}
\arguments{
\item{x}{A single-layer dNBR SpatRaster.}

\item{title}{Plot title.}

\item{subtitle}{Plot subtitle}
}
\value{
A ggplot object.
}
\description{
Plot differenced Normalized Burn Ratio
}
