% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-index.R
\name{plot_index}
\alias{plot_index}
\title{Plot a raster index}
\usage{
plot_index(
  x,
  title = NULL,
  subtitle = NULL,
  palette = sbr_palette_nbr,
  limits = NULL,
  legend_title = NULL
)
}
\arguments{
\item{x}{A single-layer terra SpatRaster.}

\item{title}{Optional plot title.}

\item{subtitle}{Optional plot subtitle.}

\item{palette}{Vector of colours.}

\item{limits}{Optional numeric vector of length two giving the
colour-scale limits.}

\item{legend_title}{Legend title.}
}
\value{
A ggplot object.
}
\description{
Plot a single-layer raster such as NBR, NDVI or dNBR.
}
