% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_index}
\alias{plot_index}
\title{Plot a continuous raster index}
\usage{
plot_index(
  x,
  index,
  title = NULL,
  subtitle = NULL,
  boundary = NULL,
  caption = NULL
)
}
\arguments{
\item{x}{A SpatRaster with one layer.}

\item{index}{Name of the registered index.}
}
\description{
Plot a continuous raster index
}
