% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_index}
\alias{plot_index}
\title{Plot a continuous raster index}
\usage{
plot_index(
  x,
  index,
  title = NULL,
  subtitle = NULL,
  boundary = NULL,
  caption = NULL
)
}
\arguments{
\item{x}{A \code{SpatRaster} containing the index to plot.}

\item{index}{Character name of the spectral index.}

\item{title}{Optional plot title.}

\item{subtitle}{Optional plot subtitle.}

\item{boundary}{Optional spatial boundary to overlay on the plot.}

\item{caption}{Optional plot caption.}
}
\description{
Plot a continuous raster index
}
