% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rgb}
\alias{plot_rgb}
\title{Plot an RGB composite}
\usage{
plot_rgb(
  x,
  title = "Sentinel-2 RGB composite",
  subtitle = NULL,
  rgb_stretch = c("lin", "hist", "none")
)
}
\arguments{
\item{x}{A composite containing red, green and blue bands.}

\item{title}{Plot title.}

\item{subtitle}{Optional subtitle.}

\item{stretch}{Apply percentile contrast stretch?}

\item{lower}{Lower percentile.}

\item{upper}{Upper percentile.}
}
\value{
A ggplot object.
}
\description{
Plot an RGB composite
}
