% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rgb.R
\name{plot_rgb}
\alias{plot_rgb}
\title{Plot RGB composite}
\usage{
plot_rgb(x, title = "RGB composite", subtitle = NULL)
}
\arguments{
\item{x}{Composite SpatRaster.}

\item{title}{Plot title.}
}
\value{
ggplot object.
}
\description{
Plot RGB composite
}
