% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_severity}
\alias{plot_severity}
\title{Plot Burn severity}
\usage{
plot_severity(
  x,
  title = "Burn severity",
  subtitle = NULL,
  caption = NULL,
  boundary = NULL
)
}
\arguments{
\item{x}{A single-layer dNBR SpatRaster.}

\item{title}{Plot title.}

\item{subtitle}{Plot subtitle}

\item{boundary}{Optional boundary to overlay on the plot.}
}
\value{
A ggplot object.
}
\description{
Plot Burn severity
}
