% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sbr_cache_clean}
\alias{sbr_cache_clean}
\title{Clean cache}
\usage{
sbr_cache_clean(temp = TRUE, downloads = FALSE, composites = FALSE)
}
\arguments{
\item{temp}{Remove temporary files.}

\item{downloads}{Remove downloaded scenes.}

\item{composites}{Remove cached sentinel-2 composites}
}
\description{
Clean cache
}
